Bug fixes
Fix issue where on logout user's information would remain in DOM
