+++ index.html
@@ -23,7 +23,7 @@
           </div>
 
         <div id="content" class="content"></div>
-
+        <button class="clear-hide btn btn-default" id="clear">Clear</button>
       <br>
       <h4 id="album-message"></h4>
       <button id="show-field-delete" class="btn btn-default">Remove an Album</button>
